Refactor password recovery module: introduce `DefaultPasswordResetController` and restructure DTOs.
