+++ src/main/java/info/mackiewicz/bankapp/system/recovery/password/service/PasswordResetService.java
@@ -1,8 +1,8 @@
 package info.mackiewicz.bankapp.system.recovery.password.service;
 
-import info.mackiewicz.bankapp.presentation.auth.dto.PasswordResetDTO;
 import info.mackiewicz.bankapp.system.notification.email.EmailService;
 import info.mackiewicz.bankapp.system.notification.email.exception.EmailSendingException;
+import info.mackiewicz.bankapp.system.recovery.password.controller.dto.PasswordResetDTO;
 import info.mackiewicz.bankapp.system.recovery.password.exception.*;
 import info.mackiewicz.bankapp.system.token.model.PasswordResetToken;
 import info.mackiewicz.bankapp.user.exception.InvalidEmailFormatException;
