+++ src/test/java/info/mackiewicz/bankapp/system/recovery/password/service/PasswordResetServiceTest.java
@@ -1,8 +1,8 @@
 package info.mackiewicz.bankapp.system.recovery.password.service;
 
-import info.mackiewicz.bankapp.presentation.auth.dto.PasswordResetDTO;
 import info.mackiewicz.bankapp.system.notification.email.EmailService;
 import info.mackiewicz.bankapp.system.notification.email.exception.EmailSendingException;
+import info.mackiewicz.bankapp.system.recovery.password.controller.dto.PasswordResetDTO;
 import info.mackiewicz.bankapp.system.recovery.password.exception.PasswordChangeException;
 import info.mackiewicz.bankapp.system.recovery.password.exception.TokenCreationException;
 import info.mackiewicz.bankapp.system.recovery.password.exception.UnexpectedTokenValidationException;
